+++ server/src/main/java/org/opensearch/action/support/replication/TransportReplicationAction.java
@@ -221,7 +221,8 @@ public abstract class TransportReplicationAction<
 
         transportService.registerRequestHandler(actionName, ThreadPool.Names.SAME, requestReader, this::handleOperationRequest);
 
-        if(transportPrimaryAction.equals(TransportShardBulkAction.ACTION_NAME + PRIMARY_ACTION_SUFFIX)){
+        // Register only TransportShardBulkAction for admission control ( primary indexing action )
+        if (transportPrimaryAction.equals(TransportShardBulkAction.ACTION_NAME + PRIMARY_ACTION_SUFFIX)) {
             transportService.registerRequestHandler(
                 transportPrimaryAction,
                 executor,