+++ server/src/main/java/org/opensearch/ratelimitting/admissioncontrol/AdmissionControlService.java
@@ -11,15 +11,13 @@ package org.opensearch.ratelimitting.admissioncontrol;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.opensearch.cluster.service.ClusterService;
-import org.opensearch.common.settings.ClusterSettings;
 import org.opensearch.common.settings.Settings;
 import org.opensearch.node.ResourceUsageCollectorService;
 import org.opensearch.ratelimitting.admissioncontrol.controllers.AdmissionController;
 import org.opensearch.ratelimitting.admissioncontrol.controllers.CPUBasedAdmissionController;
 import org.opensearch.ratelimitting.admissioncontrol.enums.AdmissionControlActionType;
 import org.opensearch.ratelimitting.admissioncontrol.stats.AdmissionControlStats;
-import org.opensearch.ratelimitting.admissioncontrol.stats.BaseAdmissionControllerStats;
-import org.opensearch.ratelimitting.admissioncontrol.stats.CPUBasedAdmissionControllerStats;
+import org.opensearch.ratelimitting.admissioncontrol.stats.AdmissionControllerStats;
 import org.opensearch.threadpool.ThreadPool;
 
 import java.util.ArrayList;
@@ -47,8 +45,14 @@ public class AdmissionControlService {
      * @param settings Immutable settings instance
      * @param clusterService ClusterService Instance
      * @param threadPool ThreadPool Instance
+     * @param resourceUsageCollectorService Instance used to get node resource usage stats
      */
-    public AdmissionControlService(Settings settings, ClusterService clusterService, ThreadPool threadPool, ResourceUsageCollectorService resourceUsageCollectorService) {
+    public AdmissionControlService(
+        Settings settings,
+        ClusterService clusterService,
+        ThreadPool threadPool,
+        ResourceUsageCollectorService resourceUsageCollectorService
+    ) {
         this.threadPool = threadPool;
         this.admissionControlSettings = new AdmissionControlSettings(clusterService.getClusterSettings(), settings);
         this.ADMISSION_CONTROLLERS = new ConcurrentHashMap<>();
@@ -68,11 +72,13 @@ public class AdmissionControlService {
 
     /**
      *
-     * @param action transport action that is being executed. we are using it for logging while request is rejected
-     * @param admissionControlActionType type of the admissionControllerActionType
+     * @param action Transport action name
+     * @param admissionControlActionType admissionControllerActionType value
      */
     public void applyTransportAdmissionControl(String action, AdmissionControlActionType admissionControlActionType) {
-        this.ADMISSION_CONTROLLERS.forEach((name, admissionController) -> { admissionController.apply(action, admissionControlActionType); });
+        this.ADMISSION_CONTROLLERS.forEach(
+            (name, admissionController) -> { admissionController.apply(action, admissionControlActionType); }
+        );
     }
 
     /**
@@ -90,7 +96,12 @@ public class AdmissionControlService {
     private AdmissionController controllerFactory(String admissionControllerName) {
         switch (admissionControllerName) {
             case CPU_BASED_ADMISSION_CONTROLLER:
-                return new CPUBasedAdmissionController(admissionControllerName, this.settings, this.clusterService, this.resourceUsageCollectorService);
+                return new CPUBasedAdmissionController(
+                    admissionControllerName,
+                    this.resourceUsageCollectorService,
+                    this.clusterService,
+                    this.settings
+                );
             default:
                 throw new IllegalArgumentException("Not Supported AdmissionController : " + admissionControllerName);
         }
@@ -113,26 +124,18 @@ public class AdmissionControlService {
         return this.ADMISSION_CONTROLLERS.getOrDefault(controllerName, null);
     }
 
-    public AdmissionControlStats stats(){
-        List<BaseAdmissionControllerStats> statsList = new ArrayList<>();
-        if(this.ADMISSION_CONTROLLERS.size() > 0){
+    /**
+     * Return admission control stats
+     */
+    public AdmissionControlStats stats() {
+        List<AdmissionControllerStats> statsList = new ArrayList<>();
+        if (this.ADMISSION_CONTROLLERS.size() > 0) {
             this.ADMISSION_CONTROLLERS.forEach((controllerName, admissionController) -> {
-                BaseAdmissionControllerStats admissionControllerStats = controllerStatsFactory(admissionController);
-                if(admissionControllerStats != null) {
-                    statsList.add(admissionControllerStats);
-                }
+                AdmissionControllerStats admissionControllerStats = new AdmissionControllerStats(admissionController, controllerName);
+                statsList.add(admissionControllerStats);
             });
             return new AdmissionControlStats(statsList);
         }
         return null;
-    }
-
-    private BaseAdmissionControllerStats controllerStatsFactory(AdmissionController admissionController) {
-        switch (admissionController.getName()) {
-            case CPU_BASED_ADMISSION_CONTROLLER:
-                return new CPUBasedAdmissionControllerStats(admissionController);
-            default:
-                return null;
-        }
     }
 }