+++ server/src/main/java/org/opensearch/ratelimitting/admissioncontrol/settings/CPUBasedAdmissionControllerSettings.java
@@ -14,9 +14,6 @@ import org.opensearch.common.settings.Settings;
 import org.opensearch.ratelimitting.admissioncontrol.AdmissionControlSettings;
 import org.opensearch.ratelimitting.admissioncontrol.enums.AdmissionControlMode;
 
-import java.util.Arrays;
-import java.util.List;
-
 /**
  * Settings related to cpu based admission controller.
  * @opensearch.internal
@@ -28,15 +25,12 @@ public class CPUBasedAdmissionControllerSettings {
      * Default parameters for the CPUBasedAdmissionControllerSettings
      */
     public static class Defaults {
-        public static final long CPU_USAGE = 95;
-        public static List<String> TRANSPORT_LAYER_DEFAULT_URI_TYPE = Arrays.asList("indexing", "search");
+        public static final long CPU_USAGE_LIMIT = 95;
     }
 
     private AdmissionControlMode transportLayerMode;
     private Long searchCPULimit;
     private Long indexingCPULimit;
-
-    private final List<String> transportActionsList;
     /**
      * Feature level setting to operate in shadow-mode or in enforced-mode. If enforced field is set
      * rejection will be performed, otherwise only rejection metrics will be populated.
@@ -54,7 +48,7 @@ public class CPUBasedAdmissionControllerSettings {
      */
     public static final Setting<Long> SEARCH_CPU_USAGE_LIMIT = Setting.longSetting(
         "admission_control.search.cpu_usage.limit",
-        Defaults.CPU_USAGE,
+        Defaults.CPU_USAGE_LIMIT,
         Setting.Property.Dynamic,
         Setting.Property.NodeScope
     );
@@ -64,7 +58,7 @@ public class CPUBasedAdmissionControllerSettings {
      */
     public static final Setting<Long> INDEXING_CPU_USAGE_LIMIT = Setting.longSetting(
         "admission_control.indexing.cpu_usage.limit",
-        Defaults.CPU_USAGE,
+        Defaults.CPU_USAGE_LIMIT,
         Setting.Property.Dynamic,
         Setting.Property.NodeScope
     );
@@ -75,7 +69,6 @@ public class CPUBasedAdmissionControllerSettings {
         clusterSettings.addSettingsUpdateConsumer(CPU_BASED_ADMISSION_CONTROLLER_TRANSPORT_LAYER_MODE, this::setTransportLayerMode);
         this.searchCPULimit = SEARCH_CPU_USAGE_LIMIT.get(settings);
         this.indexingCPULimit = INDEXING_CPU_USAGE_LIMIT.get(settings);
-        this.transportActionsList = Defaults.TRANSPORT_LAYER_DEFAULT_URI_TYPE;
         clusterSettings.addSettingsUpdateConsumer(INDEXING_CPU_USAGE_LIMIT, this::setIndexingCPULimit);
         clusterSettings.addSettingsUpdateConsumer(SEARCH_CPU_USAGE_LIMIT, this::setSearchCPULimit);
     }
@@ -102,9 +95,5 @@ public class CPUBasedAdmissionControllerSettings {
 
     public void setSearchCPULimit(Long searchCPULimit) {
         this.searchCPULimit = searchCPULimit;
-    }
-
-    public List<String> getTransportActionsList() {
-        return transportActionsList;
     }
 }